Move Emoji Panel to its own file
