+++ client/src/components/BottomPanel.tsx
@@ -1,6 +1,7 @@
-import { Drawer, IconButton, Slider, Typography } from '@material-ui/core';
+import { Drawer } from '@material-ui/core';
 import { IEmojiDict, ITowerDefenseState, PanelItemEnum } from '../types';
 import React, { useState } from 'react';
+import EmojiPanel from './EmojiPanel';
 import AnimationPanel from './AnimationPanel';
 import BackgroundPanel from './BackgroundPanel';
 import { Chat } from './Chat';
@@ -39,45 +40,6 @@ export interface ISoundPairs {
 	category: string;
 }
 
-const emojiList: IEmojiDict[] = [
-	{ name: '😍' },
-	{ name: '😎' },
-	{ name: '👏' },
-	{ name: '👀' },
-	{ name: '✨' },
-	{ name: '🎅' },
-	{
-		name: 'powerup',
-		url:
-			'https://emojis.slackmojis.com/emojis/images/1450731407/227/powerup.gif?1450731407'
-	},
-	{
-		name: 'sonic',
-		url:
-			'https://emojis.slackmojis.com/emojis/images/1450372448/149/sonic.gif?1450372448'
-	},
-	{
-		name: 'stonks',
-		url:
-			'https://emojis.slackmojis.com/emojis/images/1589323974/9036/stonks.png?1589323974'
-	},
-	{
-		name: 'bananadance',
-		url:
-			'https://emojis.slackmojis.com/emojis/images/1450694616/220/bananadance.gif?1450694616'
-	},
-	{
-		name: 'bongocat',
-		url:
-			'https://emojis.slackmojis.com/emojis/images/1537458244/4708/bongo_cat_drumming.gif?1537458244'
-	},
-	{
-		name: 'doge',
-		url:
-			'https://emojis.slackmojis.com/emojis/images/1520808873/3643/cool-doge.gif?1520808873'
-	}
-];
-
 export const BottomPanel = ({
 	bottomPanelRef,
 	isOpen,
@@ -88,50 +50,6 @@ export const BottomPanel = ({
 	updateIsTyping
 }: IPanelProps) => {
 	const [images, setImages] = useState<IImagesState[]>([]);
-	const [speed, setSpeed] = useState(1000);
-
-	const renderPanelContent = () => {
-		switch (type) {
-			case 'emoji':
-				return (
-					<>
-						{emojiList.map((emoji) => (
-							<div key={emoji.name} className="bottom-panel-emoji">
-								<IconButton
-									onClick={() => onAction('emoji', { ...emoji, speed })}
-								>
-									{emoji.url ? (
-										<img src={emoji.url} alt={emoji.name} width="24" />
-									) : (
-										emoji.name
-									)}
-								</IconButton>
-							</div>
-						))}
-						{/*<div className="bottom-panel-emoji">
-							<IconButton>...</IconButton>
-						</div>*/}
-						<div
-							style={{
-								marginLeft: '3rem',
-								marginTop: '1rem'
-							}}
-						>
-							<Typography>Speed</Typography>
-							<Slider
-								value={speed}
-								min={1000}
-								max={10000}
-								step={1000}
-								onChange={(e, value) => {
-									if (typeof value === 'number') setSpeed(value);
-								}}
-								valueLabelDisplay="auto"
-								valueLabelFormat={(x) => x / 1000}
-							/>
-						</div>
-					</>
-				);
 
 	return (
 		<Drawer variant="persistent" anchor="bottom" open={isOpen}>
@@ -162,17 +80,12 @@ const PanelContent = ({
 	switch (type) {
 		case 'emoji':
 			return (
-				<>
-					{emojiList.map((emoji) => (
-						<div key={emoji} className="bottom-panel-emoji">
-							<IconButton onClick={() => onAction('emoji', emoji)}>
-								{emoji}
-							</IconButton>
-						</div>
-					))}
-				</>
+				<EmojiPanel
+					onClick={(data: IEmojiDict) => {
+						onAction('emoji', data);
+					}}
+				/>
 			);
-
 		case 'chat':
 			return (
 				<Chat
